# nightclub-app/spec/current/openapi.yaml
openapi: 3.1.1
info:
  title: Nightclub v4 R1 Core API Design
  version: 4.0.0
  description: 開発用コア契約。未実装。ユーザー確定：名前検索中心、入口も通常承認、入口共用端末。R1全機能の詳細化状況はapi/coverage.mdを参照。実サービスURL・認証基盤・PSPは未確定。
servers:
- url: https://api.example.invalid/v1
  description: 非実在の例示URL。実接続先ではない。
security: &id001
- PersonalSession: []
- DeviceSession: []
  OperatorSession: []
paths:
...
  /stores/{storeId}/events/{eventId}/orders:
    post:
      operationId: createOrder
      tags:
      - Finance
      summary: 来店会計を作成
      description: '


        権限：sales.record。入口モードでは有効な端末・担当者・当日役割を検証。主体IDをbodyで偽装できない。'
      security: *id001
      x-permission: sales.record
      x-requirement-ids: []
      responses:
        '201':
          description: 成功（外部処理待ちは状態で明示）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderSummary'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
      parameters:
      - name: storeId
        in: path
        required: true
        schema: *id003
      - name: eventId
        in: path
        required: true
        schema: *id003
      - $ref: '#/components/parameters/CsrfToken'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/OperatorContext'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
  /stores/{storeId}/events/{eventId}/orders/{orderId}:
    get:
      operationId: getOrder
      tags:
      - Finance
      summary: 利用額・支払・未充当を表示
      description: '


        権限：sales.read。入口モードでは有効な端末・担当者・当日役割を検証。主体IDをbodyで偽装できない。'
      security: *id001
      x-permission: sales.read
      x-requirement-ids:
      - V3-PAY-05
      responses:
        '200':
          description: 成功（外部処理待ちは状態で明示）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderSummary'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
      parameters:
      - name: storeId
        in: path
        required: true
        schema: *id003
      - name: eventId
        in: path
        required: true
        schema: *id003
      - name: orderId
        in: path
        required: true
        schema: *id003
      - $ref: '#/components/parameters/OperatorContext'
...
  /stores/{storeId}/events/{eventId}/orders/{orderId}/payments:
    post:
      operationId: recordPayment
      tags:
      - Finance
      summary: 現金・外部端末の受領を記録
      description: '


        権限：payment.record。入口モードでは有効な端末・担当者・当日役割を検証。主体IDをbodyで偽装できない。'
      security: *id001
      x-permission: payment.record
      x-requirement-ids:
      - V3-PAY-07
      responses:
        '201':
          description: 成功（外部処理待ちは状態で明示）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandResult'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
      parameters:
      - name: storeId
        in: path
        required: true
        schema: *id003
      - name: eventId
        in: path
        required: true
        schema: *id003
      - name: orderId
        in: path
        required: true
        schema: *id003
      - $ref: '#/components/parameters/CsrfToken'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/OperatorContext'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRecord'
  /stores/{storeId}/events/{eventId}/payments/{paymentId}/refunds:
    post:
      operationId: createRefund
      tags:
      - Finance
      summary: 返金を冪等指示
      description: '


        権限：payment.refund。入口モードでは有効な端末・担当者・当日役割を検証。主体IDをbodyで偽装できない。'
      security: *id001
      x-permission: payment.refund
      x-requirement-ids:
      - V3-PAY-04
      responses:
        '202':
          description: 成功（外部処理待ちは状態で明示）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandResult'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
      parameters:
      - name: storeId
        in: path
        required: true
        schema: *id003
      - name: eventId
        in: path
        required: true
        schema: *id003
      - name: paymentId
        in: path
        required: true
        schema: *id003
      - $ref: '#/components/parameters/CsrfToken'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/OperatorContext'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundCreate'
...
  /stores/{storeId}/events/{eventId}/my-performance:
    get:
      operationId: getOwnPerformance
      tags:
      - Finance
      summary: 自分の集客・売上・報酬見込み
      description: '対象紹介者はサーバーが現在所属から導出。他人のmembership_idを受け取らない。報酬未合意はnull、店内未紐付けはnull。


        権限：report.own。入口モードでは有効な端末・担当者・当日役割を検証。主体IDをbodyで偽装できない。'
      security: *id001
      x-permission: report.own
      x-requirement-ids:
      - V3-FIN-03
      - V3-FIN-06
      responses:
        '200':
          description: 成功（外部処理待ちは状態で明示）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnPerformance'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
      parameters:
      - name: storeId
        in: path
        required: true
        schema: *id003
      - name: eventId
        in: path
        required: true
        schema: *id003
      - $ref: '#/components/parameters/OperatorContext'
...
components:
  securitySchemes:
    PersonalSession:
      type: apiKey
      in: cookie
      name: __Host-nc_personal
    DeviceSession:
      type: apiKey
      in: cookie
      name: __Host-nc_device
    OperatorSession:
      type: apiKey
      in: cookie
      name: __Host-nc_operator
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 160
      description: 同じ業務操作は同じキー。異なるpayloadへの流用は409。金銭は保存済み参照も照合。
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: *id002
      description: cookie認証の状態変更に必要。Origin/Fetch Metadataも検証。
    OperatorContext:
      name: X-Operator-Context
      in: header
      required: false
      schema: *id003
      description: 入口モードでは必須。画面描画時のoperator_session_id。認証手段ではない。cookie主体とサーバーactive sessionに一致しなければ409 OPERATOR_CHANGED。個人モードでは不要。
  schemas:
    Problem:
      type: object
      properties:
        type: *id002
        title: *id002
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail: *id002
        code: *id002
        trace_id: *id003
        current_version:
          type:
          - integer
          - 'null'
        retryable:
          type: boolean
        field_errors:
          type: array
          items:
            type: object
            properties:
              path: *id002
              code: *id002
              message: *id002
            required:
            - path
            - code
            - message
            additionalProperties: false
      required:
      - type
      - title
      - status
      - code
      - trace_id
      - retryable
      additionalProperties: false
...
    CommandResult:
      type: object
      properties:
        resource_id: *id003
        version: &id005
          type: integer
          minimum: 1
        status: *id002
        trace_id: *id003
        operation_key: *id002
      required:
      - resource_id
      - version
      - status
      - trace_id
      additionalProperties: false
...
    PaymentRecord:
      type: object
      properties:
        method:
          type: string
          enum:
          - CASH
          - EXTERNAL_TERMINAL
        purpose:
          type: string
          enum:
          - DEPOSIT
          - FINAL
          - ADMISSION
        amount_minor: *id008
        currency:
          type: string
          pattern: ^[A-Z]{3}$
        external_reference:
          type: string
          minLength: 1
          maxLength: 200
      required:
      - method
      - purpose
      - amount_minor
      - currency
      additionalProperties: false
    RefundCreate:
      type: object
      properties:
        amount_minor: *id008
        reason:
          type: string
          minLength: 1
          maxLength: 500
        expected_payment_version: *id005
      required:
      - amount_minor
      - reason
      - expected_payment_version
      additionalProperties: false
...
    OrderCreate:
      type: object
      properties:
        visit_id: *id003
        booking_id: *id006
        kind:
          type: string
          enum:
          - ADMISSION
          - VIP
          - IN_VENUE
        currency: *id002
      required:
      - visit_id
      - kind
      - currency
      additionalProperties: false
    OrderSummary:
      type: object
      properties:
        id: *id003
        visit_id: *id003
        version: *id005
        currency: *id002
        status: *id002
        sales_gross_minor:
          type: integer
        paid_minor: *id008
        unallocated_minor: *id008
        refund_pending_minor: *id008
      required:
      - id
      - visit_id
      - version
      - currency
      - status
      - sales_gross_minor
      - paid_minor
      - unallocated_minor
      - refund_pending_minor
      additionalProperties: false
...
    OwnPerformance:
      type: object
      properties:
        event_id: *id003
        referrer_membership_id: *id003
        registered_count:
          type: integer
          minimum: 0
        first_entries:
          type: integer
          minimum: 0
        admission_sales_minor:
          type: integer
        vip_sales_minor:
          type:
          - integer
          - 'null'
        in_venue_sales_minor:
          type:
          - integer
          - 'null'
        reward_estimate_minor:
          type:
          - integer
          - 'null'
        currency: *id002
        provisional:
          type: boolean
      required:
      - event_id
      - referrer_membership_id
      - registered_count
      - first_entries
      - admission_sales_minor
      - vip_sales_minor
      - in_venue_sales_minor
      - reward_estimate_minor
      - currency
      - provisional
      additionalProperties: false
...
  responses:
    Problem:
      description: 構造化エラー。detailsで個人情報・内部SQLを漏らさない。
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'